+++ packages/workspace/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./KnowledgeWorkspace";
-export * from "./demoVault";
 export * from "./knowledgeRoleModel";